requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
Deprecated==1.2.4
discord.py==1.0.0a1621+g66e5033
html2text==2018.1.9
idna==2.8
mccabe==0.6.1
multidict==4.5.1
pep8==1.7.1
Pillow==5.3.0
psutil==5.4.8
pycodestyle==2.4.0
pyflakes==2.0.0
PyGithub==1.43.3
PyJWT==1.7.1
requests==2.21.0
urllib3==1.24.1
websockets==6.0
wrapt==1.10.11
yarl==1.2.6
